requirements.txt
pip
Flask
cassandra-driver
requests
requests_cache
